(WIP, won't run) Overhaul device manager to generalize the core loop.
This is a checkpoint in ongoing work to decouple device_manager.py from the specific device drivers themselves. It won't get as far as having device_manager.py actually not know the device driver classes, since that would be overkill for the current state, but it's moving in that direction. A lot of stuff like pulling settings and arguments from the config data and also enumerating the timeseries definitions has been moved into the device driver classes themselves. Consolidated hubs and devices so they're all just 'devices' now, but with some parenting support.