Use real API spec & reenable chirpstack.
DeviceManager now uses the actual terraware-server APIs to communicate with it. This is not actually tested yet, but it should all be to spec now in terms of URLs and payloads.

Also: Since most testing is happening on windows machines, the build time implications of grpcio don't matter so I reenabled the Chirpstack drivers.